requirements.txt
selenium
beautifulsoup4
requests
pandas
regex
lxml
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
